+++ R/ExpressionMethods.R
@@ -23,7 +23,12 @@
 #' data(exampleTSSr)
 #' # Differential expression results are pre-computed in the @DEtables slot
 #' head(slot(exampleTSSr, "DEtables")[["control_VS_treat"]])
-setGeneric("deGene", function(object, comparePairs = list(c("control", "treat")), pval = 0.01, useMultiCore = FALSE, numCores = NULL) standardGeneric("deGene"))
+setGeneric(
+    "deGene",
+    function(object, comparePairs = list(c("control", "treat")), pval = 0.01,
+             useMultiCore = FALSE, numCores = NULL) standardGeneric("deGene"),
+    signature = "object"
+)
 #' @rdname deGene
 #' @export
 setMethod("deGene", signature(object = "TSSr"), function(object, comparePairs, pval, useMultiCore, numCores) {