+++ R/MergingMethods.R
@@ -14,7 +14,11 @@
 #' @examples
 #' data(exampleTSSr)
 #' exampleTSSr <- mergeSamples(exampleTSSr, mergeIndex = c(1, 1, 2, 2))
-setGeneric("mergeSamples", function(object, mergeIndex = NULL) standardGeneric("mergeSamples"))
+setGeneric(
+    "mergeSamples",
+    function(object, mergeIndex = NULL) standardGeneric("mergeSamples"),
+    signature = "object"
+)
 #' @rdname mergeSamples
 #' @export
 setMethod("mergeSamples", signature(object = "TSSr"), function(object, mergeIndex) {